// src/constants/enums/routerPaths.ts
export enum RouterPaths {
  PATH = '/',
  PROFILE = 'profile',
  PERSONAL_INFO = 'personal-info',
  CHANGE_PASSWORD = 'change-password',
  AUTH = 'auth',
  LOGIN = 'login',
  LOGOUT = 'logout',
  PLAN = 'plan',
  PLAN_CONFIG_COMPONENT = 'plan-config',
  PROPOSAL_MANAGEMENT = 'proposal',
  PARTNER_MANAGEMENT = 'partner',
  PARTNER_STATUS = 'status',
  PARTNER_PRODUCT = ':code/product',
  PARTNER_CONFIG = 'config',
  PARTNER_LIST_PRODUCT = 'partner-products',
  PARTNER_REMOVE_PRODUCTS = 'remove',
  LIST = 'list',
  CREATE = 'create',
  EDIT = ':id/edit',
  CREATE_CATEGORY = 'create',
  EDIT_CATEGORY = ':id/edit',
  CONFIG_CATEGORY = ':id/config',
  USER_MANAGEMENT = 'user',
  EMPLOYEE_MANAGEMENT = 'employee',
  PRODUCT_MANAGEMENT = 'product',
  CATEGORY_MANAGEMENT = 'category',
  ORDER_MANAGEMENT = 'order',
}